+++ lib/markovian/chain/compiler.rb
@@ -17,39 +17,28 @@ require 'markovian/utils/text_splitter'
 # * Handling sentences or newlines is later -- I'm not sure the right way to do it.
 # * Capitalization is deferred for later.
 module Markovian
-  class Corpus
+  class Chain
     class Compiler
       # Pass in a text, and optionally an existing Markov chain to add data to. In many cases, you
       # may be building a chain using a set of smaller texts instead of one large texts (dialog,
       # for instance, or Twitter archives), and so may call this class repeatedly for elements of
-      # the parent corpus.
-      attr_reader :corpus
-      def initialize(starter_corpus = Corpus.new)
-        @corpus = starter_corpus
+      # the corpus.
+      attr_reader :chain
+      def initialize(starter_chain = Chain.new)
+        @chain = starter_chain
       end
 
-      def build_corpus(texts)
-        texts.each {|t| incorporate_text_into_corpus(t)}
-        corpus
+      def build_chain(texts)
+        texts.each {|t| incorporate_text_into_chain(t)}
+        chain
       end
 
-      def incorporate_text_into_corpus(text)
-        add_text_to_chain(split_into_components(text), forward_chain)
-        # to assemble backward text, we just create a corpus with all the texts reversed
-        # that allows us to see what words precede any given word
-        add_text_to_chain(split_into_components(text).reverse, backward_chain)
-        corpus
+      def incorporate_text_into_chain(text)
+        add_text_to_chain(split_into_components(text), chain)
+        chain
       end
 
       protected
-
-      def forward_chain
-        corpus.forward
-      end
-
-      def backward_chain
-        corpus.backward
-      end
 
       def add_text_to_chain(text_elements, chain)
         previous_word = nil